Extract references and scripts data collecting functions
