+++ scripts/test-reviewer/createReviewPages.mjs
@@ -4,6 +4,7 @@ import { spawnSync } from 'child_process';
 import np from 'node-html-parser';
 import getReferencesData from './getReferencesData.mjs';
 import getReferenceForDirectory from './getReferenceForDirectory.mjs';
+import getScriptsData from './getScriptsData.mjs';
 import { generatePatternPages, generateIndexPage } from './generateReviewPages.mjs';
 import { commandsAPI as CommandsAPI } from '../../tests/resources/at-commands.mjs';
 
@@ -102,23 +103,11 @@ export function createReviewPages(config) {
       );
       process.exit(1);
     }
-
     referencesForPattern[directory] = referencesData;
 
-    const scriptsPath = path.join(testPlanDirectory, 'data', 'js');
-    fse.readdirSync(scriptsPath).forEach(function (scriptFile) {
-      let script = '';
-      try {
-        const data = fse.readFileSync(path.join(scriptsPath, scriptFile), 'UTF-8');
-        const lines = data.split(/\r?\n/);
-        lines.forEach(line => {
-          if (line.trim().length) script += '\t' + line.trim() + '\n';
-        });
-      } catch (err) {
-        console.error(err);
-      }
-      scripts.push(`\t${scriptFile.split('.js')[0]}: function(testPageDocument){\n${script}}`);
-    });
+    // Process test plan's various data/js/*.js script files
+    const scriptsData = getScriptsData(testPlanDirectory);
+    scripts.push(...scriptsData);
 
     fse.readdirSync(testPlanBuildDirectory).forEach(function (test) {
       if (
